Refine settings placement
