+++ apps/web/app/api/user/email-accounts/route.ts
@@ -6,10 +6,6 @@ import {
   LAST_EMAIL_ACCOUNT_COOKIE,
   parseLastEmailAccountCookieValue,
 } from "@/utils/cookies";
-import {
-  isAiSensitiveContentPolicyLocked,
-  resolveAiSensitiveContentPolicy,
-} from "@/utils/dlp/policy.server";
 import { withAuth } from "@/utils/middleware";
 
 export type GetEmailAccountsResponse = Awaited<
@@ -31,7 +27,6 @@ async function getEmailAccounts({ userId }: { userId: string }) {
       accountId: true,
       name: true,
       image: true,
-      aiSensitiveContentPolicy: true,
       account: {
         select: {
           provider: true,
@@ -50,7 +45,6 @@ async function getEmailAccounts({ userId }: { userId: string }) {
     },
   });
 
-  const policyLocked = isAiSensitiveContentPolicyLocked();
   const accountsWithRateLimits = await Promise.all(
     emailAccounts.map(async (account) => {
       const providerRateLimit = await getEmailProviderRateLimitState({
@@ -59,10 +53,6 @@ async function getEmailAccounts({ userId }: { userId: string }) {
 
       return {
         ...account,
-        aiSensitiveContentPolicy: resolveAiSensitiveContentPolicy(
-          account.aiSensitiveContentPolicy,
-        ),
-        aiSensitiveContentPolicyManaged: policyLocked,
         providerRateLimit: providerRateLimit
           ? {
               provider: providerRateLimit.provider,
